AMB-811 create only app restricted product and enforce scope on verif…

+++ tests/conftest.py
@@ -1,10 +1,6 @@
 # flake8: noqa
 import pytest
 import asyncio
-from uuid import uuid4
-from time import time
-from api_test_utils.api_test_session_config import APITestSessionConfig
-from api_test_utils.api_session_client import APISessionClient
 from api_test_utils.apigee_api_products import ApigeeApiProducts
 from api_test_utils.apigee_api_apps import ApigeeApiDeveloperApps
 from api_test_utils.oauth_helper import OauthHelper
@@ -12,52 +8,64 @@ from .configuration import config
 
 
 @pytest.fixture(scope="class")
-async def default_app():
+async def default_oauth_helper():
     """This fixture is automatically called once when used inside a class.
     The default app created here should not be modified by your tests.
     The default app has a default product associated.
     If your test requires specific app config then please create your own"""
 
-    print("\nCreating Default App and Product..")
-    apigee_product = ApigeeApiProducts()
-    await apigee_product.create_new_product()
-    await apigee_product.update_proxies(
-        [config.PROXY_NAME, "identity-service-internal-dev"]
-    )
-    await apigee_product.update_scopes(
-        ["urn:nhsd:apim:app:level3:booking-and-referral"]
-    )
+    if config.ENVIRONMENT == "int":
+        oauth = OauthHelper(config.CLIENT_ID, config.CLIENT_SECRET, config.REDIRECT_URL)
+        yield oauth
 
-    apigee_app = ApigeeApiDeveloperApps()
-    await apigee_app.create_new_app()
-
-    # Set default JWT Testing resource url
-    await apigee_app.set_custom_attributes(
-        {
-            "jwks-resource-url": "https://raw.githubusercontent.com/NHSDigital/"
-            "identity-service-jwks/main/jwks/internal-dev/"
-            "9baed6f4-1361-4a8e-8531-1f8426e3aba8.json"
-        }
-    )
-
-    await apigee_app.add_api_product(api_products=[apigee_product.name])
-
-    [
-        await product.update_ratelimits(
-            quota=60000,
-            quota_interval="1",
-            quota_time_unit="minute",
-            rate_limit="1000ps",
+    if config.ENVIRONMENT == 'internal-dev':
+        print("\nCreating Default App and Product..")
+        apigee_product = ApigeeApiProducts()
+        await apigee_product.create_new_product()
+        await apigee_product.update_proxies(
+            [config.PROXY_NAME, f"identity-service-{config.ENVIRONMENT}"]
         )
-        for product in [apigee_product]
-    ]
+        await apigee_product.update_scopes(
+            ["urn:nhsd:apim:app:level3:booking-and-referral"]
+        )
+        await apigee_product.update_environments([config.ENVIRONMENT])
 
-    yield apigee_app
+        apigee_app = ApigeeApiDeveloperApps()
+        await apigee_app.create_new_app()
 
-    # Teardown
-    print("\nDestroying Default App and Product..")
-    await apigee_app.destroy_app()
-    await apigee_product.destroy_product()
+        # Set default JWT Testing resource url
+        await apigee_app.set_custom_attributes(
+            {
+                "jwks-resource-url": "https://raw.githubusercontent.com/NHSDigital/"
+                "identity-service-jwks/main/jwks/internal-dev/"
+                "9baed6f4-1361-4a8e-8531-1f8426e3aba8.json"
+            }
+        )
+
+        await apigee_app.add_api_product(api_products=[apigee_product.name])
+
+        [
+            await product.update_ratelimits(
+                quota=60000,
+                quota_interval="1",
+                quota_time_unit="minute",
+                rate_limit="1000ps",
+            )
+            for product in [apigee_product]
+        ]
+
+        oauth = OauthHelper(
+            client_id=apigee_app.client_id,
+            client_secret=apigee_app.client_secret,
+            redirect_uri=apigee_app.callback_url,
+        )
+
+        yield oauth
+
+        # Teardown
+        print("\nDestroying Default App and Product..")
+        await apigee_app.destroy_app()
+        await apigee_product.destroy_product()
 
 
 @pytest.yield_fixture(scope="class")
@@ -68,16 +76,10 @@ def event_loop(request):
 
 
 @pytest.fixture()
-async def get_token_client_credentials(default_app):
+async def get_token_client_credentials(default_oauth_helper):
     """Call identity server to get an access token"""
-    test_app = default_app
-    oauth = OauthHelper(
-        client_id=test_app.client_id,
-        client_secret=test_app.client_secret,
-        redirect_uri=test_app.callback_url,
-    )
-    jwt = oauth.create_jwt(kid="test-1")
-    token_resp = await oauth.get_token_response(
+    jwt = default_oauth_helper.create_jwt(kid="test-1")
+    token_resp = await default_oauth_helper.get_token_response(
         grant_type="client_credentials", _jwt=jwt
     )
     return token_resp["body"]
